+++ hpc_cluster/leonhard/csv_to_SRL.py
@@ -15,7 +15,6 @@ import sys  # isort:skip
 
 sys.path.append(str(CODE_PATH))  # isort:skip
 from ipcluster import LeonhardIPCluster  # isort:skip
-from semantic_role_labeling import estimate_time, output_file  # isort:skip
 
 
 def char_length(filepath: Path, correction: int):
@@ -208,3 +207,26 @@ if __name__ == "__main__":
             raise ValueError(f"{args.srl_model_path} does not exists")
 
         run_from_batch(batch_path, SRL_MODEL_PATH)
+
+
+def estimate_time(char_length: int, device: str = "RTX2080Ti") -> float:
+    """
+    Estimate time to solution for SRL done on a given device.
+
+    """
+
+    if device == "RTX2080Ti":
+        if char_length > 10_000:
+            res = char_length * 0.3 / 1_000
+        elif char_length > 2_000:
+            res = char_length * 0.6 / 1_000
+        else:
+            res = 1.0
+    else:
+        raise ValueError("{device} not estimated.")
+
+    return res
+
+
+def output_file(filepath: pathlib.Path, parent_path: pathlib.Path) -> pathlib.Path:
+    return (parent_path / filepath.name).with_suffix(".json")
